+++ src/serializer/statistics.js
@@ -31,6 +31,7 @@ export class SerializerStatistics extends RealmStatistics {
     this.checkThatFunctionsAreIndependent = new PerformanceTracker(getTime, getMemory);
     this.processCollectedNestedOptimizedFunctions = new PerformanceTracker(getTime, getMemory);
     this.deepTraversal = new PerformanceTracker(getTime, getMemory);
+    this.referentialization = new PerformanceTracker(getTime, getMemory);
     this.referenceCounts = new PerformanceTracker(getTime, getMemory);
     this.serializePass = new PerformanceTracker(getTime, getMemory);
     this.babelGenerate = new PerformanceTracker(getTime, getMemory);
@@ -99,6 +100,7 @@ export class SerializerStatistics extends RealmStatistics {
   processCollectedNestedOptimizedFunctions: PerformanceTracker;
   deepTraversal: PerformanceTracker;
   referenceCounts: PerformanceTracker;
+  referentialization: PerformanceTracker;
   serializePass: PerformanceTracker;
   babelGenerate: PerformanceTracker;
   dumpIR: PerformanceTracker;
@@ -142,10 +144,10 @@ export class SerializerStatistics extends RealmStatistics {
     );
     console.log(
       `${format(this.deepTraversal)} visiting residual heap, ${format(
-        this.referenceCounts
-      )} reference counting, ${format(this.serializePass)} generating AST, ${format(
-        this.babelGenerate
-      )} generating source code`
+        this.referentialization
+      )} referentializing functions, ${format(this.referenceCounts)} reference counting, ${format(
+        this.serializePass
+      )} generating AST, ${format(this.babelGenerate)} generating source code`
     );
   }
 }
